CG-133 Refactor: 바텀탭네비게이터 코드 분기
TODO: Nesting Navigator 구현 전락 세워야 함
https://reactnavigation.org/docs/nesting-navigators/
